+++ src/components/import/feature/importObject.ts
@@ -1,28 +1,21 @@
 // importObject.ts
 import csvToJson from "@/fileImportServices/csvServices"
 
-// Updated function to accept state setters as arguments
-export function handleFilesSelected(
-  selectedFiles: File[],
-  setFiles: React.Dispatch<React.SetStateAction<File[]>>,
-  setCsvData: React.Dispatch<React.SetStateAction<any[]>>
-) {
-  setFiles(selectedFiles)
-
-  if (selectedFiles.length > 0) {
-    const file = selectedFiles[0]
+export function parseCsvFile(file: File): Promise<any[]> {
+  return new Promise((resolve, reject) => {
     const reader = new FileReader()
 
     reader.onload = (event) => {
       const csvString = event.target?.result as string
       const parsedData = csvToJson(csvString)
-      setCsvData(parsedData)
+      resolve(parsedData)
     }
 
     reader.onerror = () => {
       console.error("Error reading file")
+      reject("Error reading file")
     }
 
     reader.readAsText(file) // Read the file as text
-  }
+  })
 }
