+++ src/components/import/layout/dragNDropLayout.tsx
@@ -1,59 +1,42 @@
-// File: DragNDropLayout.tsx
-import React, { useEffect, useState, DragEvent, ChangeEvent } from "react"
-import DragNdrop from "../ui/dragNDrop"
+// /components/import/layout/dragNDropLayout.tsx
+import React from "react"
+import DragNdrop from "../feature/dragNDrop"
+import PreviewTable from "../ui/previewTable"
 
-// Define the prop types for the component
 interface DragNDropLayoutProps {
-  onFilesSelected: (files: File[]) => void
   supportedFileTypes?: string
   className?: string
+  csvData: any[]
+  setCsvData: React.Dispatch<React.SetStateAction<any[]>>
+  setFiles: React.Dispatch<React.SetStateAction<File[]>>
+  files: File[]
 }
 
 const DragNDropLayout: React.FC<DragNDropLayoutProps> = ({
-  onFilesSelected,
   supportedFileTypes = ".pdf,.docx,.pptx,.txt,.xlsx", // Default supported types
   className,
+  csvData,
+  setCsvData,
+  setFiles,
+  files,
 }) => {
-  const [files, setFiles] = useState<File[]>([])
-
-  // Handle file selection from input
-  const handleFileChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const selectedFiles = event.target.files
-    if (selectedFiles && selectedFiles.length > 0) {
-      const newFiles = Array.from(selectedFiles)
-      setFiles((prevFiles) => [...prevFiles, ...newFiles])
-    }
-  }
-
-  // Handle file drop
-  const handleDrop = (event: DragEvent<HTMLDivElement>) => {
-    event.preventDefault()
-    const droppedFiles = event.dataTransfer.files
-    if (droppedFiles.length > 0) {
-      const newFiles = Array.from(droppedFiles)
-      setFiles((prevFiles) => [...prevFiles, ...newFiles])
-    }
-  }
-
-  // Handle file removal
-  const handleRemoveFile = (index: number) => {
-    setFiles((prevFiles) => prevFiles.filter((_, i) => i !== index))
-  }
-
-  useEffect(() => {
-    onFilesSelected(files)
-  }, [files, onFilesSelected])
-
   return (
-    <DragNdrop
-      onFilesSelected={(files) => setFiles(files)}
-      handleFileChange={handleFileChange}
-      handleDrop={handleDrop}
-      handleRemoveFile={handleRemoveFile}
-      files={files}
-      supportedFileTypes={supportedFileTypes}
-      className={className}
-    />
+    <div>
+      {csvData.length === 0 ? (
+        <DragNdrop
+          supportedFileTypes={supportedFileTypes}
+          className={className}
+          csvData={csvData}
+          setCsvData={setCsvData}
+          setFiles={setFiles}
+          files={files}
+        />
+      ) : (
+        <section className="mt-4">
+          <PreviewTable data={csvData} />
+        </section>
+      )}
+    </div>
   )
 }
